==> Rcal/01_RDA_invertebrate_calc.R <==
# 01_RDA_invertebrate_calc
# written by: sean Kinard
# last edit: 2023-05-19
#----------------------------------------------------------------------------
# Description
#----------------------------------------------------------------------------
# The script conducts Redundancy Analysis (RDA) to explore the relationship between invertebrate community composition and environmental variables, generating coordinates, axis information, and fitted vectors, and exporting the results for further analysis.

#----------------------------------------------------------------------------
# Setup
#----------------------------------------------------------------------------
source('Rcal/00_setup_packages.R')

dc <- read_csv("Rdat/communities.csv") %>%
  rename(site_code = site, collection_date = date)

de <- read_csv("Rdat/RDA_environment.csv")

#----------------------------------------------------------------------------
# Merge to community
#----------------------------------------------------------------------------
df <- de %>%
  select(-annual_rainfall) %>%
  right_join(dc %>% 
               filter(guild =='invertebrate') %>%
               group_by(site_code, collection_date, month, lowest_taxon) %>%
               summarize(abun = sum(abun)) %>%
               pivot_wider(names_from = lowest_taxon, values_from = abun) %>%
               mutate(UID = paste(site_code,collection_date, sep = '_')) %>%
               select(site_code, collection_date, month, UID, Argia:Odontomyia) %>%
               replace(is.na(.), 0) %>%
               ungroup()) %>%
  unique() %>%
  select(site_code, collection_date, everything())

#----------------------------------------------------------------------------
# Final Matrices for RDA
#----------------------------------------------------------------------------
spec <- df %>%
  column_to_rownames("UID") %>%
  select(c(Argia:Odontomyia))
env2 <- df %>%
  column_to_rownames("UID") %>%
  select(flow_2wk_mean:canopy) %>%
  rename(Conductivity = conductivity,
         Flow = flow_2wk_mean,
         Nitrate = nitrate,
         "Max Depth" = depth_max,
         Algae = algae,
         Canopy = canopy)

#----------------------------------------------------------------------------
# Redundancy Analysis (RDA) 
#----------------------------------------------------------------------------
# step 1: The ordiplot object is obtained from the result of via function rda. The ordination is done with a community data set that is transformed by the Hellinger method.

# script generated by the BiodiversityR GUI from the constrained ordination menu
invertebrate.Hellinger <- disttransform(spec, method='hellinger')
Ordination.model2 <- rda(invertebrate.Hellinger ~ . ,data=env2, scaling="species")

# This can be plotted via ordiplot
plot2 <- ordiplot(Ordination.model2, choices=c(1,2))

# Extract Site Coordinates
RDA_coordinates_site <- sites.long(plot2, env.data=env2) %>%
  separate(labels, into=c('site', 'collection_date'), sep='_') %>%
  mutate(site = factor(site, levels = c("Semi-Arid", "Transition", "Sub-Humid"))) %>%
  mutate(collection_date = ymd(collection_date),
         x = collection_date) %>%
  separate(x, into = c('year', 'month', 'day'), sep='-') %>%
  mutate(month=substr(month,2,2)) %>%
  select(site, collection_date, year, month, day, axis1, axis2) %>%
  tibble() %>%
  rename(labels=site) %>%
  mutate(c_type='Site')

# Extract Species Coordinates
RDA_coordinates_species <- species.long(plot2) %>%
  tibble() %>%
  mutate(c_type='Species')

# Information on the labeling of the axes is obtained with function axis.long. This information is extracted from the ordination model and not the ordiplot, hence it is important to select the same axes via argument 'choices'. The extracted information includes information on the percentage of explained variation. Be advised to cross-check with the summary of the redundancy analysis, where information on proportion explained is given.

# Merge Coordinates for export
RDA_coordinates_out <- full_join(RDA_coordinates_site, RDA_coordinates_species)

# Extract Variance per Axis
RDA_axis_out <- axis.long(Ordination.model2, choices=c(1, 2))

#----------------------------------------------------------------------------
# Fit Vectors
#----------------------------------------------------------------------------
# Species
spec.envfit <- envfit(plot2, env=invertebrate.Hellinger)
spec.data.envfit <- data.frame(r2=spec.envfit$vectors$r, p=spec.envfit$vectors$pvals)
RDA_species_vector <- species.long(plot2, spec.data=spec.data.envfit) %>% tibble() %>%
  rename(vector=labels) %>%
  mutate(v_type='Species')

# Environment
env.envfit <- envfit(plot2, env= env2)
RDA_env_vector <- vectorfit.long(env.envfit) %>%
  tibble() %>%
  rename(r2=r) %>%
  mutate(v_type='Physical')

# Merge fitted vectors for export
RDA_vectors <- full_join(RDA_species_vector,RDA_env_vector)

#----------------------------------------------------------------------------
# Export
#----------------------------------------------------------------------------
write_csv(RDA_coordinates_out, 'Rout/RDA_coordinates_invertebrate.csv')
write_csv(RDA_axis_out, 'Rout/RDA_axis_invertebrate.csv')
write_csv(RDA_vectors, 'Rout/RDA_vectors_invertebrate.csv')

#----------------------------------------------------------------------------
# End 01_RDA_invertebrate_calc
